features: Make Main Package Feature Tests Public
Move all feature teats in the main package to a new
internal directory, "internal/features", which can be used
by the main package and also mirrored in the "features"
package for public use.
